requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.34.0

# Database
SQLAlchemy==2.0.45
PyMySQL==1.1.0
mysql-connector-python==9.5.0
alembic==1.17.2

# Data Validation
pydantic[email]==2.12.5
pydantic-settings==2.7.0

# Environment Variables
python-dotenv==1.0.1

# HTTP Client (外部API呼び出し用)
httpx==0.28.1
requests==2.32.3

# Cache
cachetools==5.5.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pyjwt==2.10.1
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# Scheduler
APScheduler==3.10.4

# Azure Services
azure-identity==1.19.0
openai==1.59.5

# Testing (開発時のみ)
pytest==8.3.4
pytest-asyncio==0.25.2

# Code Quality
ruff==0.9.1

# Email
resend
email-validator
